improvements to screenshots test suite
