+++ L10nSnapshotTests/L10nSuite1SnapshotTests.swift
@@ -3,8 +3,6 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
 import XCTest
-let testPageBase2 = "http://www.example.com"
-let loremIpsumURL2 = "\(testPageBase2)"
 
 class L10nSuite1SnapshotTests: L10nBaseSnapshotTests {
     override var skipIntro: Bool {
@@ -20,7 +18,7 @@ class L10nSuite1SnapshotTests: L10nBaseSnapshotTests {
         }
     }
     
-    func test1WebViewContextMenu() {
+    func testWebViewContextMenu() {
         // Drag the context menu up to show all the options
         func drag() {
             let window = XCUIApplication().windows.element(boundBy: 0)
@@ -31,8 +29,8 @@ class L10nSuite1SnapshotTests: L10nBaseSnapshotTests {
 
         // Link
         navigator.openURL("http://wikipedia.org")
-        waitUntilPageLoad()
         waitForNoExistence(app.staticTexts["XCUITests-Runner pasted from Fennec"])
+        waitForExistence(app.webViews.element(boundBy: 0).links.element(boundBy: 0), timeout: 5)
         navigator.goto(WebLinkContextMenu)
         drag()
         snapshot("WebViewContextMenu-01-link")
@@ -40,86 +38,89 @@ class L10nSuite1SnapshotTests: L10nBaseSnapshotTests {
 
         // Image
         navigator.openURL("http://wikipedia.org")
+        waitForNoExistence(app.staticTexts["XCUITests-Runner pasted from Fennec"])
+        waitForExistence(app.webViews.element(boundBy: 0).images.element(boundBy: 0), timeout: 5)
         navigator.goto(WebImageContextMenu)
         drag()
         snapshot("WebViewContextMenu-02-image")
         navigator.back()
     }
 
-    func test2WebViewAuthenticationDialog() {
+    func testWebViewAuthenticationDialog() {
         navigator.openURL("https://jigsaw.w3.org/HTTP/Basic/", waitForLoading: false)
         waitForNoExistence(app.staticTexts["XCUITests-Runner pasted from Fennec"])
         navigator.goto(BasicAuthDialog)
         snapshot("WebViewAuthenticationDialog-01", waitForLoadingIndicator: false)
         navigator.back()
     }
-    // Disabled due to real bug:  https://github.com/mozilla-mobile/firefox-ios/issues/7248
-    /*func test3ReloadButtonContextMenu() {
-        navigator.toggleOn(userState.trackingProtectionSettingOnNormalMode == true, withAction: Action.SwitchETP)
-        navigator.goto(BrowserTab)
+//    // Disabled due to real bug:  https://github.com/mozilla-mobile/firefox-ios/issues/7248
+//    /*func test3ReloadButtonContextMenu() {
+//        navigator.toggleOn(userState.trackingProtectionSettingOnNormalMode == true, withAction: Action.SwitchETP)
+//        navigator.goto(BrowserTab)
+//
+//        navigator.openURL(loremIpsumURL2)
+//        navigator.toggleOff(userState.requestDesktopSite, withAction: Action.ToggleRequestDesktopSite)
+//        navigator.goto(ReloadLongPressMenu)
+//        snapshot("ContextMenuReloadButton-01")
+//        navigator.toggleOn(userState.requestDesktopSite, withAction: Action.ToggleRequestDesktopSite)
+//        navigator.goto(ReloadLongPressMenu)
+//        snapshot("ContextMenuReloadButton-02", waitForLoadingIndicator: false)
+//
+//        navigator.toggleOff(userState.trackingProtectionPerTabEnabled, withAction: Action.ToggleTrackingProtectionPerTabEnabled)
+//        navigator.goto(ReloadLongPressMenu)
+//
+//        // Snapshot of 'Reload *with* tracking protection' label, because trackingProtectionPerTabEnabled is false.
+//        snapshot("ContextMenuReloadButton-03", waitForLoadingIndicator: false)
+//    }*/
+//
+//    private func typePasscode(n: Int, keyNumber: Int) {
+//        for _ in 1...n {
+//            app.keys.element(boundBy: keyNumber).tap()
+//            sleep(1)
+//        }
+//    }
+//
+//    func test4PasscodeSettings() {
+//        navigator.goto(PasscodeSettings)
+//        app.tables.cells["TurnOnPasscode"].tap()
+//        snapshot("SetPasscodeScreen-1-nopasscode")
+//        
+//        // Type "111111 passcode"
+//        typePasscode(n: 6, keyNumber: 2)
+//        snapshot("SetPasscodeScreen-2-typepasscode")
+//        // Type incorrect passcode "111112"
+//        typePasscode(n: 5, keyNumber: 2)
+//        // Type once inkey "2"
+//        typePasscode(n: 1, keyNumber: 1)
+//        snapshot("SetPasscodeScreen-3-passcodesmustmatch")
+//        
+//        // Confitm passcode
+//        typePasscode(n: 6, keyNumber: 2)
+//        typePasscode(n: 6, keyNumber: 2)
+//        snapshot("SetPasscodeScreen-3")
+//        
+//        // Go to interval settings
+//        app.tables.cells["PasscodeInterval"].tap()
+//        typePasscode(n: 6, keyNumber: 2)
+//        snapshot("PasscodeIntervalScreen-1")
+//    }
 
-        navigator.openURL(loremIpsumURL2)
-        navigator.toggleOff(userState.requestDesktopSite, withAction: Action.ToggleRequestDesktopSite)
-        navigator.goto(ReloadLongPressMenu)
-        snapshot("ContextMenuReloadButton-01")
-        navigator.toggleOn(userState.requestDesktopSite, withAction: Action.ToggleRequestDesktopSite)
-        navigator.goto(ReloadLongPressMenu)
-        snapshot("ContextMenuReloadButton-02", waitForLoadingIndicator: false)
-
-        navigator.toggleOff(userState.trackingProtectionPerTabEnabled, withAction: Action.ToggleTrackingProtectionPerTabEnabled)
-        navigator.goto(ReloadLongPressMenu)
-
-        // Snapshot of 'Reload *with* tracking protection' label, because trackingProtectionPerTabEnabled is false.
-        snapshot("ContextMenuReloadButton-03", waitForLoadingIndicator: false)
-    }*/
-
-    private func typePasscode(n: Int, keyNumber: Int) {
-        for _ in 1...n {
-            app.keys.element(boundBy: keyNumber).tap()
-            sleep(1)
-        }
-    }
-
-    func test4PasscodeSettings() {
-        navigator.goto(PasscodeSettings)
-        app.tables.cells["TurnOnPasscode"].tap()
-        snapshot("SetPasscodeScreen-1-nopasscode")
-        
-        // Type "111111 passcode"
-        typePasscode(n: 6, keyNumber: 2)
-        snapshot("SetPasscodeScreen-2-typepasscode")
-        // Type incorrect passcode "111112"
-        typePasscode(n: 5, keyNumber: 2)
-        // Type once inkey "2"
-        typePasscode(n: 1, keyNumber: 1)
-        snapshot("SetPasscodeScreen-3-passcodesmustmatch")
-        
-        // Confitm passcode
-        typePasscode(n: 6, keyNumber: 2)
-        typePasscode(n: 6, keyNumber: 2)
-        snapshot("SetPasscodeScreen-3")
-        
-        // Go to interval settings
-        app.tables.cells["PasscodeInterval"].tap()
-        typePasscode(n: 6, keyNumber: 2)
-        snapshot("PasscodeIntervalScreen-1")
-    }
-
-    func test5TopSitesMenu() {
+    func testTopSitesMenu() {
         navigator.goto(HomePanel_TopSites)
         navigator.goto(TopSitesPanelContextMenu)
         snapshot("TopSitesMenu-01")
     }
 
-    func test6HistoryTableContextMenu() {
-        navigator.openURL(loremIpsumURL2)
+    func testHistoryTableContextMenu() {
+        navigator.openURL(loremIpsumURL)
         waitForNoExistence(app.staticTexts["XCUITests-Runner pasted from Fennec"])
+        
         navigator.goto(HistoryPanelContextMenu)
         snapshot("HistoryTableContextMenu-01")
     }
 
-    func test7BookmarksTableContextMenu() {
-        navigator.openURL(loremIpsumURL2)
+    func testBookmarksTableContextMenu() {
+        navigator.openURL(loremIpsumURL)
         // There is no other way the test work with the new Copied.. snackbar ahow on iOS14
         waitForNoExistence(app.staticTexts["XCUITests-Runner pasted from Fennec"])
         waitForExistence(app.buttons["TabLocationView.pageOptionsButton"], timeout: 5)
@@ -141,7 +142,7 @@ class L10nSuite1SnapshotTests: L10nBaseSnapshotTests {
         snapshot("21ReaderModeSettingsMenu-01")
     }*/
 
-    func test8ETPperSite() {
+    func testETPperSite() {
         // Enable Strict ETP
         navigator.goto(TrackingProtectionSettings)
         // Check the warning alert
@@ -151,7 +152,7 @@ class L10nSuite1SnapshotTests: L10nBaseSnapshotTests {
          app.alerts.buttons.firstMatch.tap()
 
         // Website without blocked elements
-        navigator.openURL(loremIpsumURL2)
+        navigator.openURL(loremIpsumURL)
         waitForNoExistence(app.staticTexts["XCUITests-Runner pasted from Fennec"])
         waitForExistence(app.buttons["TabLocationView.trackingProtectionButton"], timeout: 5)
         navigator.goto(TrackingProtectionContextMenuDetails)
@@ -162,7 +163,7 @@ class L10nSuite1SnapshotTests: L10nBaseSnapshotTests {
         snapshot("TrackingProtectionDisabledPerSite-02")
     }
 
-    func test9SettingsETP() {
+    func testSettingsETP() {
         navigator.goto(TrackingProtectionSettings)
 
         waitForExistence(app.cells["Settings.TrackingProtectionOption.BlockListBasic"])
