+++ L10nSnapshotTests/L10nSuite2SnapshotTests.swift
@@ -4,69 +4,84 @@
 
 import XCTest
 
-let testPageBase = "http://www.example.com"
-let loremIpsumURL = "\(testPageBase)"
+//let testPageBase = "http://www.example.com"
+//let loremIpsumURL = "\(testPageBase)"
 
 class L10nSuite2SnapshotTests: L10nBaseSnapshotTests {
-    func test1DefaultTopSites() {
-        navigator.toggleOff(userState.pocketInNewTab, withAction: Action.TogglePocketInNewTab)
-        navigator.goto(HomePanelsScreen)
-        snapshot("DefaultTopSites-01")
-        navigator.toggleOn(userState.pocketInNewTab, withAction: Action.TogglePocketInNewTab)
-        navigator.goto(HomePanelsScreen)
-        snapshot("DefaultTopSites-with-pocket-02")
-    }
 
-    func test2MenuOnTopSites() {
-        navigator.goto(NewTabScreen)
-        navigator.goto(BrowserTab)
-        waitUntilPageLoad()
-        navigator.goto(BrowserTabMenu)
-        snapshot("MenuOnTopSites-01")
-    }
-
-    func test3Settings() {
-        let table = app.tables.element(boundBy: 0)
-        navigator.goto(SettingsScreen)
-        table.forEachScreen { i in
-            snapshot("Settings-main-\(i)")
-        }
-
-        allSettingsScreens.forEach { nodeName in
-            self.navigator.goto(nodeName)
-            table.forEachScreen { i in
-                snapshot("Settings-\(nodeName)-\(i)")
-            }
-        }
-    }
-
-    func test4PrivateBrowsingTabsEmptyState() {
-        navigator.toggleOn(userState.isPrivate, withAction: Action.TogglePrivateMode)
-        snapshot("PrivateBrowsingTabsEmptyState-01")
-    }
-
-    func test5PanelsEmptyState() {
+    func testPanelsEmptyState() {
         let libraryPanels = [
-            "LibraryPanels.History",
-            "LibraryPanels.ReadingList",
-            "LibraryPanels.Downloads",
-            "LibraryPanels.SyncedTabs"
-        ]
-        navigator.goto(LibraryPanel_Bookmarks)
-        snapshot("PanelsEmptyState-LibraryPanels.Bookmarks")
-        libraryPanels.forEach { panel in
-            app.buttons[panel].tap()
-            snapshot("PanelsEmptyState-\(panel)")
-        }
+                "LibraryPanels.History",
+                "LibraryPanels.ReadingList",
+                "LibraryPanels.Downloads",
+                "LibraryPanels.SyncedTabs"
+            ]
+            navigator.goto(LibraryPanel_Bookmarks)
+            snapshot("PanelsEmptyState-LibraryPanels.Bookmarks")
+            libraryPanels.forEach { panel in
+                app.buttons[panel].tap()
+                snapshot("PanelsEmptyState-\(panel)")
+            }
     }
+
+//    func test1DefaultTopSites() {
+//        navigator.toggleOff(userState.pocketInNewTab, withAction: Action.TogglePocketInNewTab)
+//        navigator.goto(HomePanelsScreen)
+//        snapshot("DefaultTopSites-01")
+//        navigator.toggleOn(userState.pocketInNewTab, withAction: Action.TogglePocketInNewTab)
+//        navigator.goto(HomePanelsScreen)
+//        snapshot("DefaultTopSites-with-pocket-02")
+//    }
+//
+//    func test2MenuOnTopSites() {
+//        navigator.goto(NewTabScreen)
+//        navigator.goto(BrowserTabMenu)
+//        snapshot("MenuOnTopSites-01")
+//    }
+//
+//    func test3Settings() {
+//        let table = app.tables.element(boundBy: 0)
+//        navigator.goto(SettingsScreen)
+//        table.forEachScreen { i in
+//            snapshot("Settings-main-\(i)")
+//        }
+//
+//        allSettingsScreens.forEach { nodeName in
+//            self.navigator.goto(nodeName)
+//            table.forEachScreen { i in
+//                snapshot("Settings-\(nodeName)-\(i)")
+//            }
+//        }
+//    }
+//
+//    func test4PrivateBrowsingTabsEmptyState() {
+//        navigator.toggleOn(userState.isPrivate, withAction: Action.TogglePrivateMode)
+//        snapshot("PrivateBrowsingTabsEmptyState-01")
+//    }
+//
+//    func test5PanelsEmptyState() {
+//        let libraryPanels = [
+//            "LibraryPanels.History",
+//            "LibraryPanels.ReadingList",
+//            "LibraryPanels.Downloads",
+//            "LibraryPanels.SyncedTabs"
+//        ]
+//        navigator.goto(LibraryPanel_Bookmarks)
+//        snapshot("PanelsEmptyState-LibraryPanels.Bookmarks")
+//        libraryPanels.forEach { panel in
+//            app.buttons[panel].tap()
+//            snapshot("PanelsEmptyState-\(panel)")
+//        }
+//    }
 
     // From here on it is fine to load pages
-    func test6LongPressOnTextOptions() {
+    func testLongPressOnTextOptions() {
         navigator.openURL(loremIpsumURL)
         waitUntilPageLoad()
         waitForNoExistence(app.staticTexts["XCUITests-Runner pasted from Fennec"])
 
         // Select some text and long press to find the option
+        waitForExistence(app.webViews.element(boundBy: 0).staticTexts.element(boundBy: 0), timeout: 10)
         app.webViews.element(boundBy: 0).staticTexts.element(boundBy: 0).press(forDuration: 1)
         snapshot("LongPressTextOptions-01")
         if(app.menuItems["show.next.items.menu.button"].exists) {
@@ -75,7 +90,7 @@ class L10nSuite2SnapshotTests: L10nBaseSnapshotTests {
         }
     }
 
-    func test7URLBar() {
+    func testURLBar() {
         navigator.goto(URLBarOpen)
         snapshot("URLBar-01")
 
@@ -84,7 +99,7 @@ class L10nSuite2SnapshotTests: L10nBaseSnapshotTests {
         snapshot("URLBar-02")
     }
 
-    func test8URLBarContextMenu() {
+    func testURLBarContextMenu() {
         // Long press with nothing on the clipboard
         navigator.goto(URLBarLongPressMenu)
         snapshot("LocationBarContextMenu-01-no-url")
@@ -96,7 +111,7 @@ class L10nSuite2SnapshotTests: L10nBaseSnapshotTests {
         snapshot("LocationBarContextMenu-02-with-url")
     }
 
-    func test10MenuOnWebPage() {
+    func testMenuOnWebPage() {
         navigator.openURL(loremIpsumURL)
         waitForNoExistence(app.staticTexts["XCUITests-Runner pasted from Fennec"])
         navigator.goto(BrowserTabMenu)
@@ -110,16 +125,16 @@ class L10nSuite2SnapshotTests: L10nBaseSnapshotTests {
         navigator.back()
     }
 
-    func test10PageMenuOnWebPage() {
+    func testPageMenuOnWebPage() {
         navigator.goto(BrowserTab)
         waitForNoExistence(app.staticTexts["XCUITests-Runner pasted from Fennec"])
         waitForExistence(app.buttons["TabLocationView.pageOptionsButton"], timeout: 15)
         navigator.goto(PageOptionsMenu)
         snapshot("MenuOnWebPage-03")
         navigator.back()
-       }
+    }
 
-    func test11FxASignInPage() {
+    func testFxASignInPage() {
         navigator.goto(Intro_FxASignin)
         waitForExistence(app.navigationBars.staticTexts["FxASingin.navBar"])
         snapshot("FxASignInScreen-01")
